Simplify the provider API by making suitable() also prepare providers
There was not much point in having a separate prepare() call for a
provider, especially since it was supposed to be called after we checked
for suitability, which in practice was pointless.

Also changed suitable() from returning just a bool to a result<bool> to
make it possible to report errors during provider preparation.

Finally, cleaned up the logic in ral::types a little bit to ensure we do
use the same logic when setting up providers and creating types from them.