result.hpp: change ok() and err() to throw if accessed 'the wrong way'
Trying to call ok() on a result that is not ok, or err() on a result that
is ok will now throw a std::logic_error, rather than returning boost::none.

Using boost::optional here was too careful and annoying, as users of result
should always check whether the result is ok or not separately before
accessing it.