Identify providers as TYPE::PROVIDER instead of just by type
We do not have a system in place to select a default provider for a
type. Therefore, make identifying providers by a qualified name
TYPE::PROVIDER the primary way to identify providers. As a convenience,
ralsh also accepts just TYPE if there is only one provider for that type.

Types currently are simply a string, there is no other logic tied to it,
and they are therefore more a hint to users than anything functional.