(DI-136) Add/update scripts used by Go binding
Takes the version to install as an argument
```
./install_go.sh 1.8.3
```

Prior to this the build.sh script would always attempt to clone the libral repo, this commit updates the script to first check if there is already a libral dir present at `/usr/src/libral` which will result in the clone being skipped.

It also clears any existing build directory before creating it.

Adds a build script, based on the existing static libral build script,
for the simple example go client `examples/go/listproviders.go`.

This is intended to run within the libral-build-go container.

Adds a script to shrink the build example client binary (upx and strip),
and package up the libral external dependencies (providers/augeas lenses
etc).