(DI-136) Use published image names and add go image to makefile
This commit switches to using the published `libral/el6-build-static`
image name which is pushed to the hub. It also uses a publishable name
for the go variant so that we can pull libral go build images from the
hub.

It also updates the libral/el6-build-static-go image to skip pulling
in scripts which will have already been copied to the base
libral/el6-build-static image, and adds a make target for the go
image.