(MAINT) Fix absent handling for Python2.7+
This prior to this commit calling do_get in the DNF provider would fail
due to the attempt to concatenate a list with a view.

This commit converts the view to a list for the concatenation and also
corrects the absent package handling so that absent packages are only
returned when explicitly querying a package name.