(DI-362) Expose ability to SetResource in Go binding
This commit exposes the set_resource function via the c wrapper to allow
resources to be created/updated via the Go binding.

This commit adds a new Change type which describes the changes to a
resource after calling `SetResource`.

It also removes the `Raw` attribute from the Resource type which was
unused and redundant (it contained a raw []byte representaion of the
resource).

This commit adds the SetResource function to the libralgo package which
exposes the ability to set attributes on a resource.

Refer to the function godoc for more information.