Change travis setup so that we use the git checkout from Travis
We used to do our own git checkout of the code, which was wrong, as we
would always run against master, even when we were testing a PR or other
branch.

We now map Travis' git checkout into our container. We also use the
build.sh file from the checkout so that changing that doesn't require
rebuilding the containers.

Fixes: https://github.com/puppetlabs/libral/issues/51