(DI-834) Update find_state for systemd so it works with older versions
The find_state function was relying on the ability to pass the unit file
to `systemctl list-unit-files` command, the ability to do so was added
at some point after systemd 208. RHEL shipped with version 208, so
attempting to get the status of a service on an un-upgraded system would
subsequently fail.

This commit switches to using systemctl with grep in the find_state
function to determine the state of a service.