(DI-867) Systemd provider returns all units (except instanced)
This commit updates the join to deal with units reported by
`list-unit-files` but subsequently not reported by `list-units` by
setting them to a default state.

This appears to work well for all but instanced services such as getty@
which appear as the abstract unit rather than an instance.

At present support for instanced services is lacking in the provider
*and* in Puppet itself, so for the moment the main instanced service
will be omitted from results (see issue #65).

The difference in behaviours has been discussed at length here:
https://bugzilla.redhat.com/show_bug.cgi?id=748512